Exports+Maxoid Updates
Updated the Export process to write directly to source files rather than into clipboard in order to make compiling new maze collections faster and less error prone.
Also, added encoding for the 'Number of HIts' to kill a Maxoid which was actually configurable but always being set to 4 (this seemed high), so now the Editor allows a single sparkoid hit to kill if needed.