Fix: Change the way we handle redirected stdout/stderror with MAME
MAME outputs a lot of stderror text for the setup we have and we need to asynchronously parse it as it comes in or the stream will hang the HBMAME process. Async processing is code I don't what to deal with ATM. Ignore this for now. At some point I can look back and pull in the MS example code.