Attempt at fixing ImageList issue (failed) but changes are good
So, I tried to fix this issue I was having with the imagelist throwing an exception when loading the Menu Item Resource Images... thought it was that crappy .ico file in there so I replaced that with a decent .png... but that wasn't it. I also tried commenting out all the image adds in the GameExplorer constructor, but it was still throwing an exception..  so more research needs to be done... thought it was a corrupt image, but they al seem okay?