Fixed Resource Loading Issue
Turns out that when I changed the namespace of the MajorHavocEditor project back to 'mhedit', that it broke all the resources. I had attempted to change the Namespace as that is what drives the ClickOnce application. Im not going to type out why that is important here, but suffice to say, I changed the Namespace back to MajorHavocEditor for now as it is critical to compilation. We can figure out how to best use the ClickOnce platform using a new namespace for this UX version of the editor.